Resolves #55 - Parser fails on large files due to arithmetic overflow.
