+++ lib/saxWasm.js
@@ -1,6 +1,5 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.SAXParser = exports.Tag = exports.Text = exports.ProcInst = exports.Attribute = exports.AttributeType = exports.Position = exports.Reader = exports.SaxEventType = void 0;
 class SaxEventType {
 }
 exports.SaxEventType = SaxEventType;
@@ -183,7 +182,7 @@ exports.Tag = Tag;
 class SAXParser {
     constructor(events = 0, options = { highWaterMark: 32 * 1024 }) {
         this.eventTrap = (event, ptr, len) => {
-            const uint8array = new Uint8Array(this.wasmSaxParser.memory.buffer, ptr, len);
+            const uint8array = new Uint8Array(this.wasmSaxParser.memory.buffer, ptr, len).slice();
             let detail;
             switch (event) {
                 case SaxEventType.Attribute:
@@ -221,8 +220,7 @@ class SAXParser {
                     if (self.wasmSaxParser) {
                         self.wasmSaxParser.parser(events);
                     }
-                },
-                configurable: false, enumerable: true
+                }, configurable: false, enumerable: true
             }
         });
     }
@@ -239,7 +237,7 @@ class SAXParser {
         if (!this.writeBuffer || this.writeBuffer.buffer !== memory.buffer) {
             this.writeBuffer = new Uint8Array(memory.buffer);
         }
-        this.writeBuffer.set(chunk);
+        this.writeBuffer.set(chunk, 0);
         write(0, chunk.byteLength);
     }
     end() {
