+++ package.json
@@ -38,7 +38,7 @@
     "libxmljs": "^0.19.7",
     "ltx": "^2.9.2",
     "mocha": "^9.2.0",
-    "node-expat": "^2.3.18",
+    "node-expat": "^2.4.0",
     "node-xml": "^1.0.2",
     "nyc": "^15.1.0",
     "sax": "^1.2.4",
